Add model file
Contains the classes to which the network wil be parsed. At this
time, the network is fully parsed into the network classes, which
contains other subclasses.

The __init__ function for the network takes as input the XML path,
and the subclasses take as input dictionaries parsed by python's
ElementTree library.